LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.rotate.nonfinite.worker-expected.txt


PASS rotate() with Infinity/NaN is ignored
